+++ source/SoloReadBarcode_getCBandUMI.cpp
@@ -31,7 +31,7 @@ void SoloReadBarcode::getCBandUMI(string &readNameExtra)
     //convert CB and check for Ns
     int32 posN=convertNuclStrToInt32(cbSeq,cbB);
 
-    if (pSolo.cbWL.size()==0) {//no whitelist - no search
+    if (!pSolo.cbWLyes) {//no whitelist - no search
         if (posN!=-1) {//Ns are present, discard this read
             stats.V[stats.nNinBarcode]++;
         } else {//no Ns